+++ CHANGELOG.md
@@ -18,6 +18,19 @@ The changelog for **SwifterSwiftUI**. Also see the [releases](https://github.com
 
 ---
 
+## [v0.3.2](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.3.2)
+
+### Added
+- **Result**
+    - Result extensions to simplify access to success and failure values.
+- **UserDefaultEnum**
+    - Property wrapper to use the user's defaults with enum type with string raw value.
+- **UserDefault**
+    - Property wrapper to use the user's defaults with default types.
+
+### Changed
+- Rename **Storage** property wrapper to **UserDefault**.
+
 ## [v0.3.1](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.3.1)
 
 ### Added